Move interface generator to own module
